fix(garcom/contextos_de_negocio/identidade_e_acesso/pontos_de_entrada): move arquivo de modelos pydantic para pasta
