+++ garcom/contextos_de_negocio/identidade_e_acesso/pontos_de_entrada/modelos_pydantic/representacoes.py
